Asaxiy_Scraper/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
decouple==0.0.4
Django==5.0
django-clearcache==1.2.1
idna==3.6
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
schedule==1.2.1
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0

